services/simcct/requirements.txt
bcrypt==3.1.7
celery[redis]==4.3.0
coverage==4.5.3
elastic-apm[flask]==5.1.2
email-validator==1.0.4
Flask-Bcrypt==0.7.1
flask-caching==1.7.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-Testing==0.7.1
fluent-logger==0.9.3
geoip2==2.9.0
gunicorn==19.9.0
mongoengine==0.18.2
prettytable==0.7.2
PyJWT==1.7.1
pymongo==3.9.0
python-dateutil==2.8.0
redis==3.2.1
